[FIX] Fix duplicated ping issue
The issue was caused by copying the dma structure in the driver's
`eth_rx_complete` function.

In this case function `client_rx_data()` was returning to the dma pool
the pointer to the copy instead of to the original, what at some point
(full `pending_rx` ring buffer cycle) was causing referencing wrong dma
buffer.

Duplicated ping is just one of the symptoms of this general RX issue.

Fixes:
SEOS-2134