[FIX] SEOS-3041: Give all available RX buffers to the client
This fixes the reoccurring RX buffer overflow errors that were
especially present when the driver was configured with the 
promiscuous mode ON in an environment with high network 
traffic. The root cause of this issue was related the various buffers
that the driver manages to provide the incoming data to the 
connected client component. The initial driver had support for
several connected clients and therefore had the clients share
the available RX buffers. Since our driver was reduced (for now) to
only support one connected client, we can give all available RX
buffers to this client.